Relocate factory support to its own namespace
